Exercise 1 and 2. Code rewritten to cope with large expansion factor.

+++ day11/src/main/java/se/nyquist/Point.java
@@ -1,6 +1,6 @@
 package se.nyquist;
 
-public record Point(int x, int y) {
+public record Point(long x, long y) {
     public static int compare(Point p1, Point p2) {
         var yCompare = Long.compare(p1.y,p2.y);
         return yCompare == 0 ? Long.compare(p1.x,p2.x) : yCompare;
@@ -10,11 +10,11 @@ public record Point(int x, int y) {
         return new Point(this.x + delta.x, this.y + delta.y);
     }
 
-    public Point addY(int y) {
+    public Point addY(long y) {
         return new Point(this.x, this.y + y);
     }
 
-    public Point addX(int x) {
+    public Point addX(long x) {
         return new Point(this.x + x, this.y);
     }
 
@@ -22,7 +22,7 @@ public record Point(int x, int y) {
         return new Point(this.x - start.x, this.y - start.y);
     }
 
-    public Point extend(int i) {
+    public Point extend(long i) {
         return new Point(this.x + i, this.y + i);
     }
 }
